A few changes to improve compatibility

+++ src/UIBubbleTableView.m
@@ -95,7 +95,7 @@
 	self.bubbleSection = nil;
     
     // Loading new data
-    int count = 0;
+    NSUInteger count = 0;
 #if !__has_feature(objc_arc)
     self.bubbleSection = [[[NSMutableArray alloc] init] autorelease];
 #else
@@ -156,7 +156,7 @@
 
 - (NSInteger)numberOfSectionsInTableView:(UITableView *)tableView
 {
-    int result = [self.bubbleSection count];
+    NSInteger result = [self.bubbleSection count];
     if (self.typingBubble != NSBubbleTypingTypeNobody) result++;
     return result;
 }
@@ -169,7 +169,7 @@
     return [[self.bubbleSection objectAtIndex:section] count] + 1;
 }
 
-- (float)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath
+- (CGFloat)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath
 {
     // Now typing
 	if (indexPath.section >= [self.bubbleSection count])
